feat(api): 接口类分家,Facade 只留五个任务方法，轨道库与 spatiography 各自成类（#610）
按 ADR 0043：
- 新增 e2m2e.api.catalog.Catalog（七个轨道库方法 + orbit_family_generation，
  auto_ingest/load_record_ephemeris 为 Facade 任务方法的入库接缝）与
  e2m2e.api.spatiography.Spatiography（五个分区分析工具）
- Facade 变组合根：exposed_apis 给出扫描根，catalog/spatiography 属性
  交出实例；二档占位声明移除（落地时进领域类，不回 Facade）
- tool_inventory/tool_spec/tool_specs/CLI 子命令派生跨类解析属主，
  清单仍单一来源；工具名与 18 个工具面不变，MCP/CLI/sidecar 零感知
- 测试与示例调用点迁移到新类；placeholder 过滤机制改由桩对象验证
